requirements.txt
# Core AI and ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=10.0.0

# Web Framework and UI
streamlit>=1.28.0
streamlit-webrtc>=0.45.0

# Mobile App Framework
kivy>=2.1.0
kivymd>=1.1.0
plyer>=2.1.0

# APK Building
buildozer>=1.5.0
cython>=0.29.0

# Image Processing
opencv-python>=4.8.0
matplotlib>=3.6.0

# Data Handling
pandas>=1.5.0
scipy>=1.10.0

# Utilities
requests>=2.28.0
tqdm>=4.64.0
python-dateutil>=2.8.0

# Development and Testing
pytest>=7.2.0
black>=22.10.0
flake8>=5.0.0

# Security and Privacy
cryptography>=3.4.8

# Android Permissions (for mobile app)
# Note: These are installed automatically by buildozer
# pyjnius>=1.4.2
# android>=0.1.0
